requirements.txt
diffusers
transformers
einops
huggingface_hub
opencv-python
trimesh
omegaconf
scikit-image
numpy==1.22.3
peft
jaxtyping
typeguard
diso
torch-cluster